+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - [ ] Indication of a rules that are already defined by parent .editorconfig
 - [ ] Indication of rules not matching any files
 - [ ] Validate syntax of sections
+- [x] Rich QuickInfo tooltips
+- [x] Rich completion tooltips
 - [x] Dark theme support
 - [x] Support for AnyCode
 - [x] Show Intellisense when severity glyph is clicked